Added get_user and checks
Created get_user function that gets the user and returns a ShowUser
schema response, as to not reveal the password. Created a check to see
if the user or document exists in the collection. If it does, it will
send a True boolean.

Hid the URI for the DB in case migration to a cloud MongoDB server.