Added get_all_users and updated documentation
In the user_handlers.py, using asynchronous list comprehension to get
all users in the users collection.

In the main.py, description and necessary details have been added to
give more information about the API in the documentation system.

In the users.py, proper docstrings has been added to provide "guidance"
how to use the API.