Added user authentication and su functionality
auth_handlers.py
The function fetches either of the user's email or username, if none is
fetched then it will raise an error. If it does, it will check for the
validity of the user through verification of the password input. If not
the same password, it will raise an error.

By default, the token expiration is 30 minutes. Giving enough time to do
most things in the API. It creates an encoded JWT using a function to be
made later.

oauth2.py
Provides the get_current_user dependency to make sure the user still has
their token.

token.py
Handles every function that is needed for the JWT Authentication. It
handles the verification of the token and creation.

checks.py
Added a class called RoleChecker that is used as a dependency for role
based access control. It will take a list of roles and checks the user's
role if it is in the list, if not, access will be denied.

users.py
Added the su access for every path.

main.py
Fixed the documentation description