Added upload to library route
main.py
Added "documentation" for library routes

hashing.py
Added a functionality to hash files in preparation for vt-py
implementation for virus detection.

checks.py
Ignore changes, trying to refactor the check_if_exists function but
failed.

library.py
The routing file used in the main.py for include_router.

lib_handlers.py
Checks if the book title already exists in the database, if it has the
same filename, if the file has the same hash. If checks are passed,
upload file to GridFS then return schemas.ShowBook

schemas.py
Made a Pydantic model for Book and ShowBook, had to explicitly make a
dependency inside the book to allow "JSON" or Form data and Upload file
at the same time.

db.py
Made it so that is in each wrapper and to make dependencies easier.
GridFS cannot be used as dependency.